refactor: split app page stage content
